move Table to database_content.py
