+++ postgres_declare/mixins/grantable.py
@@ -3,16 +3,14 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Sequence
 
-from sqlalchemy import Inspector, TextClause, inspect, text
+from sqlalchemy import TextClause, text
 
 from postgres_declare.data_structures.grant_to import GrantTo
 from postgres_declare.data_structures.privileges import Privilege
 from postgres_declare.entities.entity import Entity
 from postgres_declare.exceptions import EntityExistsError, InvalidPrivilegeError
-from postgres_declare.mixins.sql import SQLBase
 
 if TYPE_CHECKING:
-    from postgres_declare.entities.database_content import DatabaseContent
     from postgres_declare.entities.role import Role
 
 
@@ -109,57 +107,3 @@ class GrantableEntity(Grantable, Entity):
     def grant(self, grants: Sequence[GrantTo]) -> None:
         super().grant(grants=grants)
         self._fix_entity_order(grants=grants, target_entity=self)
-
-
-class Table(SQLBase, Grantable):
-    def __init__(self, name: str, database_content: DatabaseContent, schema: str | None = "public"):
-        super().__init__(name=name)
-        self.database_content = database_content
-        self.schema = schema
-
-    def __hash__(self) -> int:
-        return hash((self.name, self.__class__.__name__, self.database_content.name, self.schema))
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, self.__class__):
-            return NotImplemented
-        return (self.name, self.__class__.__name__, self.database_content.name, self.schema) == (
-            other.name,
-            other.__class__.__name__,
-            other.database_content.name,
-            self.schema,
-        )
-
-    def grant(self, grants: Sequence[GrantTo]) -> None:
-        super().grant(grants=grants)
-        self._fix_entity_order(grants=grants, target_entity=self.database_content)
-
-    def _exists(self) -> bool:
-        # TODO not sure how expensive creating an inspector is, might not want to do it for every run of this fn
-        inspector: Inspector = inspect(self.database_content.database.db_engine())
-        return inspector.has_table(table_name=self.name, schema=self.schema)
-
-    def _grant(self, grantee: Role, privileges: set[Privilege]) -> None:
-        self._commit_sql(
-            engine=self.database_content.database.db_engine(),
-            statements=self._grant_statements(grantee=grantee, privileges=privileges),
-        )
-
-    def _revoke(self, grantee: Role, privileges: set[Privilege]) -> None:
-        self._commit_sql(
-            engine=self.database_content.database.db_engine(),
-            statements=self._revoke_statements(grantee=grantee, privileges=privileges),
-        )
-
-    @staticmethod
-    def _allowed_privileges() -> set[Privilege]:
-        return {
-            Privilege.INSERT,
-            Privilege.SELECT,
-            Privilege.UPDATE,
-            Privilege.DELETE,
-            Privilege.TRUNCATE,
-            Privilege.REFERENCES,
-            Privilege.TRIGGER,
-            Privilege.ALL_PRIVILEGES,
-        }
